cmd_context_select v1.2
Changed the logic again to avoid situation when pressing the button doesn't change anything.

If cursor is almost directly next to the unit it will select all units from the same group.
If unit is not in group it will select all visible units of this type.
Otherwise, it will select units close to the cursor. If selection didn't change it will go to the next step:
 1. Units with relative hp (default <50%).
 2. One builder. Repeating will cycle the builders in range.
 3. One skuttle/spybot, close to the cursor
 4. All radar and jammers, close to the cursor
 5. Otherwise - cycle your labs across the whole map.